README.md
This is a mission for APC members of the GTA community to enjoy. I made this simply for fun and i hope other enjoy this mission. All scripting that I have done will be credited to me, and any other authors that took part will too be credited. 

If you are wanting to borrow this mission you must: 

- Be granted permission by myself 
- Do NOT change the author's, if it is made by someone else and not you, don't change them
- You may change map features in the mission file
- Do NOT change the onLoadMission name. It is to stay the same, as is the author

This mission is only for Altis. No mods are required for this mission, but if you are wanting to add in mods once you have permission to use this you may.